+++ client/src/components/GridTemplateContainer.jsx
