+++ client/src/pages/SignUp.jsx
@@ -7,14 +7,12 @@ import {
   TextField,
   FormHelperText,
 } from "@material-ui/core";
-import LoginSignupContainer from "pages/LoginSignupContainer";
-import GridTemplateContainer from "pages/GridTemplateContainer";
+import LoginSignupContainer from "components/LoginSignupContainer";
+import GridTemplateContainer from "components/GridTemplateContainer";
+import Onboarding from "./Onboarding";
 
 // TODO Figure out where to move useStyles to avoid duplicate code
 const useStyles = makeStyles((theme) => ({
-  grid: {
-    flexGrow: 1,
-  },
   form: {
     "& .MuiTextField-root": {
       margin: theme.spacing(1),
@@ -26,7 +24,7 @@ const useStyles = makeStyles((theme) => ({
     alignItems: "center",
   },
   button: {
-    marginTop: 30,
+    marginTop: 15,
     marginBottom: 60,
     borderRadius: 20,
     color: "#FFFFFF",
@@ -48,7 +46,6 @@ const useStyles = makeStyles((theme) => ({
   },
   h1: {
     fontSize: "xx-large",
-    marginBottom: 50,
   },
   formHelper: {
     marginBottom: 10,
@@ -56,51 +53,84 @@ const useStyles = makeStyles((theme) => ({
   },
 }));
 
-const Login = () => {
+const SignUp = () => {
   const classes = useStyles();
+  const [userName, setUserName] = useState("");
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
+  const [confirmedPass, setConfirmedPass] = useState("");
   const [error, setError] = useState("");
+  const [nextPage, setNextPage] = useState(false);
 
   const handleChange = (event) => {
     const { name, value } = event.target;
 
     // eslint-disable-next-line
     switch (name) {
+      case "userName":
+        setUserName(value);
+        break;
       case "email":
         setEmail(value);
         break;
       case "password":
         setPassword(value);
         break;
+      case "confirmedPass":
+        setConfirmedPass(value);
+        break;
+      case "error":
+        setError(value);
+        break;
     }
   };
 
   const submit = async (event) => {
     event.preventDefault();
-    try {
-      const { data } = await axios.post("/api/users/login", {
-        email: email,
-        password: password,
-      });
 
-      localStorage.token = data.token;
+    if (password.length < 7) {
+      setError("Password needs to be at least 7 characters long");
+    } else if (password === confirmedPass) {
+      try {
+        const { data } = await axios.post("/api/users/signup", {
+          name: userName,
+          email: email,
+          password: password,
+        });
 
-      if (localStorage.token) {
-        // TODO redirect to homepage
-        console.log(localStorage.token);
+        localStorage.token = data.token;
+
+        if (localStorage.token) {
+          setNextPage(true);
+        }
+      } catch (err) {
+        console.error(err);
+        setError(err.response.data.response);
       }
-    } catch (err) {
-      console.error(err);
-      setError(err.response.data.response);
+    } else {
+      setError("Passwords do not match");
     }
   };
 
-  return (
+  return nextPage ? (
+    <Onboarding />
+  ) : (
     <LoginSignupContainer>
       <GridTemplateContainer>
         <form onSubmit={submit} className={classes.form}>
-          <h1 className={classes.h1}>Welcome back!</h1>
+          <h1 className={classes.h1}>Create an account</h1>
+          <TextField
+            id="userName-input"
+            type="text"
+            label="Name"
+            name="userName"
+            variant="outlined"
+            required
+            value={userName}
+            onChange={handleChange}
+            className={classes.textfield}
+          />
+
           <TextField
             id="email-input"
             type="email"
@@ -125,6 +155,18 @@ const Login = () => {
             className={classes.textfield}
           />
 
+          <TextField
+            id="confirmedPass-input"
+            type="password"
+            label="Confirm Password"
+            name="confirmedPass"
+            variant="outlined"
+            required
+            value={confirmedPass}
+            onChange={handleChange}
+            className={classes.textfield}
+          />
+
           {error && (
             <FormHelperText className={classes.formHelper}>
               {error}
@@ -132,14 +174,14 @@ const Login = () => {
           )}
 
           <Button type="submit" variant="contained" className={classes.button}>
-            Login
+            Continue
           </Button>
 
           <div>
             <strong>
-              Don't have an account?{" "}
-              <Link to="/sign-up" className={classes.link}>
-                Create
+              Already have an account?{" "}
+              <Link to="/login" className={classes.link}>
+                Login
               </Link>
             </strong>
           </div>
@@ -149,4 +191,4 @@ const Login = () => {
   );
 };
 
-export default Login;
+export default SignUp;
